+++ resources/views/lead/form.blade.php
@@ -52,20 +52,15 @@
         {{ Form::remark('lead[remark]') }}
         {{
           Form::survey('lead[survey]', [
-            's2' => '2D Seismic',
-            'geo' => 'Geological',
-            'chem' => 'Geochemistry',
-            'grav' => 'Gravity',
-            'elec' => 'Electromagnetic',
-            'resi' => 'Resistivity',
-            'oter' => 'Other'
+            'lead[s2_data]' => '2D Seismic',
+            'lead[geo_data]' => 'Geological',
+            'lead[chem_data]' => 'Geochemistry',
+            'lead[grav_data]' => 'Gravity',
+            'lead[elec_data]' => 'Electromagnetic',
+            'lead[resi_data]' => 'Resistivity',
+            'lead[oter_data]' => 'Other'
           ])
         }}
-
-        {{ Form::text('lead[waaaaaa]', 'fuu', ['id' => 'wa']) }}
-        @push('jsready')
-          $("#wa").hide();
-        @endpush
 
         @if (actionName() === 'edit')
           {{ Form::bsTextarea('lead[update_reason]', 'Update reason', true) }}
@@ -81,9 +76,7 @@
       </div>
       <div class="panel-body">
 
-        {{-- 2D Seismic Placeholder --}}
-        <span id="s2-place" style="display: none;"></span>
-        <div id="s2-panel" class="panel panel-default">
+        <div id="s2-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               2D Seismic
@@ -112,9 +105,7 @@
           </div>
         </div> {{-- 2D Seismic --}}
 
-        {{-- Geological Field Placeholder --}}
-        <span id="geo-place" style="display: none;"></span>
-        <div id="geo-panel" class="panel panel-default">
+        <div id="geo-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               Geological Field
@@ -135,9 +126,7 @@
           </div>
         </div> {{-- Geological Field --}}
 
-        {{-- Geochemistry Placeholder --}}
-        <span id="chem-place" style="display: none;"></span>
-        <div id="chem-panel" class="panel panel-default">
+        <div id="chem-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               Geochemistry
@@ -156,9 +145,7 @@
           </div>
         </div> {{-- Geochemistry --}}
 
-        {{-- Gravity Placeholder --}}
-        <span id="grav-place" style="display: none;"></span>
-        <div id="grav-panel" class="panel panel-default">
+        <div id="grav-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               Gravity
@@ -182,9 +169,7 @@
           </div>
         </div> {{-- Gravity --}}
 
-        {{-- Electromagnetic Placeholder --}}
-        <span id="elec-place" style="display: none;"></span>
-        <div id="elec-panel" class="panel panel-default">
+        <div id="elec-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               Electromagnetic
@@ -208,9 +193,7 @@
           </div>
         </div> {{-- Electromagnetic --}}
 
-        {{-- Resistivity Placeholder --}}
-        <span id="resi-place" style="display: none;"></span>
-        <div id="resi-panel" class="panel panel-default">
+        <div id="resi-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               Resistivity
@@ -234,9 +217,7 @@
           </div>
         </div> {{-- Resistivity --}}
 
-        {{-- Other Placeholder --}}
-        <span id="oter-place" style="display: none;"></span>
-        <div id="oter-panel" class="panel panel-default">
+        <div id="oter-panel" class="panel panel-default hidden">
           <div class="panel-heading">
             <div class="panel-title">
               Other